Fix data type. pkg/empleo-9A645583F8F24B28

+++ pkg/empleo-9A645583F8F24B28/upgrade.sql
@@ -5,7 +5,7 @@ CREATE OR REPLACE TABLE empleo (
     asignacion_salarial VARCHAR(50),
     codigo_empleo VARCHAR(10), -- assigned by simo
     sin_codigo BOOL,
-    denominacion_id INT,
+    denominacion_id TINYINT,
     grado_nivel JSON,
     grado_denominacion JSON,
     convocatoria_id INT,
